requirements.txt
anyio==4.4.0
certifi==2024.6.2
charset-normalizer==3.3.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
praw==7.7.1
prawcore==2.4.0
PySide6==6.7.2
PySide6_Addons==6.7.2
PySide6_Essentials==6.7.2
python-dateutil==2.9.0.post0
requests==2.32.3
shiboken6==6.7.2
six==1.16.0
sniffio==1.3.1
update-checker==0.18.0
urllib3==2.2.2
websocket-client==1.8.0
